+++ R/additive_make.R
@@ -28,11 +28,7 @@ additive_make <- function(modes = c("classification", "regression")) {
     # -------------------------------------------------------------------------
 
     for (pkg in dependpkgs) {
-      if ("mode" %in% rlang::fn_fmls_names(parsnip::set_dependency)) {
-        parsnip::set_dependency(model, engine, pkg = pkg, mode = mode)
-      } else {
-        parsnip::set_dependency(model, engine, pkg = pkg)
-      }
+      parsnip::set_dependency(model, engine, pkg = pkg, mode = mode)
     }
 
     # -------------------------------------------------------------------------
@@ -202,15 +198,6 @@ additive_make <- function(modes = c("classification", "regression")) {
     parsnip::set_model_arg(
       model = model,
       eng = engine,
-      parsnip = "weights",
-      original = "weights",
-      func = fitfunc,
-      has_submodel = FALSE
-    )
-
-    parsnip::set_model_arg(
-      model = model,
-      eng = engine,
       parsnip = "subset",
       original = "subset",
       func = fitfunc,
@@ -351,7 +338,7 @@ additive_make <- function(modes = c("classification", "regression")) {
       mode = mode,
       value = list(
         interface = "formula",
-        protect = c("formula", "data"),
+        protect = c("formula", "data", "weights"),
         func = fitfunc,
         defaults = list()
       )
@@ -382,11 +369,14 @@ additive_make <- function(modes = c("classification", "regression")) {
         value = list(
           pre = NULL,
           post = function(results, object) {
+            threshold <- getOption("class_pred.threshold", 0.5)
             if (length(object$lvl) == 2) {
               if (is.array(results)) {
                 results <- as.vector(results)
               }
-              threshold <- getOption("class_pred.threshold", 0.5)
+              if (length(threshold) != 1) {
+                rlang::abort("Probability threshold should be a single value.")
+              }
               if (is.numeric(threshold)) {
                 if (!dplyr::between(threshold, 0, 1)) {
                   rlang::abort("Probability threshold is out of 0-1 range.")
@@ -403,6 +393,9 @@ additive_make <- function(modes = c("classification", "regression")) {
               length(object$lvl) > 2 &
                 length(object$lvl) == ncol(results)
             ) {
+              if (length(threshold) == ncol(results)) {
+                results <- sweep(results, 2, threshold, FUN = "/")
+              }
               results <- object$lvl[apply(results, 1, which.max)]
             } else {
               rlang::abort("Unexpected model predictions!")
@@ -427,6 +420,9 @@ additive_make <- function(modes = c("classification", "regression")) {
           pre = NULL,
           post = function(results, object) {
             if (length(object$lvl) == 2) {
+              if (is.array(results)) {
+                results <- as.vector(results)
+              }
               results <- tibble::tibble(
                 v1 = 1 - results,
                 v2 = results